requirements.txt
Django==1.11.10
mysqlclient==1.3.12
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
